requirements.txt
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

sae-bench
eai-sparsify==1.3.0
transformers==4.49.0
datasets==3.2.0
accelerate==1.2.1
safetensors==0.5.3

# #torch
# #adapt to your cuda drivers: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
# transformers
# datasets
# accelerate
# safetensors
# pyyaml

# eai-sparsify
# sae-bench
# sentence-transformers

# #requires linux/ WSL: git+https://github.com/EleutherAI/delphi.git